support embed images in libpurple
Parse <img> tags in messages, store the image in the filesystem and
replace the tag with a url to the image.

This add two new settings, webserver_root and webserver_url, and only
stores images if they exist.

The user should configure an http server, pointing to the webserver_root
and serving to webserver_url

Example Config:
webserver_url = http://example.com
webserver_root = /srv/bitlbee/